src/scrapers/awsdeploy/requirements.txt
beautifulsoup4==4.9.3
boto3==1.15.16
botocore==1.18.16
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
DateTime==4.3
idna==2.10
jmespath==0.10.0
lxml==4.5.2
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
zope.interface==5.1.2
